# api/src/modules/doctor/doctor.controller.ts
import {
  Body,
  Controller,
  Get,
  Param,
  Post,
  Put,
  UseGuards,
  Query,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiOperation,
  ApiParam,
  ApiResponse,
  ApiTags,
  ApiQuery,
} from '@nestjs/swagger';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { RolesGuard } from '../auth/guards/roles.guard';
import { Roles } from '../../common/decorators/roles.decorator';
import { UserRole } from '../user/types/role.enum';
import { DoctorService } from './doctor.service';
import { CreateDoctorDto } from './dto/create-doctor.dto';
import { UpdateDoctorDto } from './dto/update-doctor.dto';
import { TopDoctorsQueryDto } from './dto/top-doctors.dto';
import { DoctorParamsDto } from './dto/doctor-params.dto';
import { Doctor } from './entities/doctor.entity';

@ApiTags('Врачи')
@Controller('doctors')
export class DoctorController {
  constructor(private readonly doctorService: DoctorService) {}

  @Post()
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(UserRole.SUPER_ADMIN)
  @ApiBearerAuth('access-token')
  @ApiBody({ type: CreateDoctorDto })
  @ApiResponse({
    status: 200,
    description: 'Успешно создан',
  })
  async create(@Body() createDto: CreateDoctorDto): Promise<Doctor> {
    return this.doctorService.create(createDto);
  }

  @Get('top')
  @ApiOperation({
    summary: 'Получить топ докторов',
    description:
      'Получение списка лучших докторов, отсортированных по рейтингу от лучшего к худшему',
  })
  @ApiQuery({
    name: 'limit',
    description: 'Количество докторов для возврата (1-100)',
    example: 10,
    required: false,
  })
  @ApiQuery({
    name: 'minRating',
    description: 'Минимальный рейтинг для фильтрации (1.0-5.0)',
    example: 4.0,
    required: false,
  })
  @ApiQuery({
    name: 'specialization',
    description: 'Фильтр по специализации',
    example: 'хирург',
    required: false,
  })
  @ApiQuery({
    name: 'clinicId',
    description: 'ID клиники для фильтрации',
    example: 1,
    required: false,
  })
  @ApiResponse({
    status: 200,
    description: 'Список топ докторов с рейтингами',
    type: [Doctor],
  })
  async getTopDoctors(@Query() query: TopDoctorsQueryDto): Promise<Doctor[]> {
    return this.doctorService.getTopDoctors(query);
  }

  @Get('search')
  @ApiOperation({
    summary: 'Поиск врачей',
    description: 'Поиск врачей по имени, специализации и описанию',
  })
  @ApiQuery({
    name: 'query',
    description: 'Поисковый запрос',
    example: 'кардиолог',
    required: true,
  })
  @ApiResponse({
    status: 200,
    description: 'Список найденных врачей',
    type: [Doctor],
  })
  async searchDoctors(@Query('query') query: string): Promise<Doctor[]> {
    return this.doctorService.searchDoctors(query);
  }

  @Get('search/specialization')
  @ApiOperation({
    summary: 'Поиск врачей по специализации',
    description: 'Поиск врачей по конкретной специализации',
  })
  @ApiQuery({
    name: 'specialization',
    description: 'Специализация',
    example: 'хирург',
    required: true,
  })
  @ApiResponse({
    status: 200,
    description: 'Список врачей с указанной специализацией',
    type: [Doctor],
  })
  async searchBySpecialization(
    @Query('specialization') specialization: string,
  ): Promise<Doctor[]> {
    return this.doctorService.searchDoctorsBySpecialization(specialization);
  }

  @Put(':id')
  @ApiBearerAuth('access-token')
  @UseGuards(JwtAuthGuard, RolesGuard)
  @Roles(UserRole.SUPER_ADMIN)
  @ApiOperation({
    summary: 'Обновить информацию о враче',
    description:
      'Обновление информации о враче (только для супер-администраторов)',
  })
  @ApiParam({
    name: 'id',
    description: 'ID врача',
    type: Number,
  })
  @ApiBody({ type: UpdateDoctorDto })
  @ApiResponse({
    status: 200,
    description: 'Информация о враче успешно обновлена',
  })
  @ApiResponse({
    status: 400,
    description: 'Некорректный ID врача',
  })
  @ApiResponse({
    status: 404,
    description: 'Врач не найден',
  })
  async update(
    @Param() params: DoctorParamsDto,
    @Body() updateDto: UpdateDoctorDto,
  ): Promise<Doctor> {
    return this.doctorService.update(params.id, updateDto);
  }

  @Get(':id')
  @ApiOperation({
    summary: 'Получить информацию о враче',
    description: 'Получение полной информации о враче',
  })
  @ApiParam({
    name: 'id',
    description: 'ID врача',
    type: Number,
  })
  @ApiResponse({
    status: 200,
    description: 'Информация о враче',
  })
  @ApiResponse({
    status: 400,
    description: 'Некорректный ID врача',
  })
  @ApiResponse({
    status: 404,
    description: 'Врач не найден',
  })
  async findOne(@Param() params: DoctorParamsDto): Promise<Doctor> {
    return this.doctorService.findOne(params.id);
  }

  @Get(':id/schedule')
  @ApiOperation({
    summary: 'Получить врача с расписанием',
    description: 'Полная информация о враче вместе с его расписанием',
  })
  @ApiParam({
    name: 'id',
    description: 'ID врача',
    type: Number,
  })
  @ApiResponse({
    status: 200,
    description: 'Информация о враче и его расписании',
  })
  @ApiResponse({
    status: 400,
    description: 'Некорректный ID врача',
  })
  @ApiResponse({
    status: 404,
    description: 'Врач не найден',
  })
  async getDoctorSchedule(@Param() params: DoctorParamsDto) {
    return this.doctorService.getDoctorWithSchedule(params.id);
  }
}