# ui/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/admin/update-role": {
      "post": {
        "operationId": "AdminController_updateUserRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Роль успешно обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeUserDto"
                }
              }
            }
          },
          "403": {
            "description": "Запрещено"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновление роли",
        "tags": ["Admin"]
      }
    },
    "/user/register": {
      "post": {
        "operationId": "UserController_completeRegistration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Телефон не подтверждён"
          }
        },
        "summary": "Регистрация пользовтеля",
        "tags": ["User"]
      }
    },
    "/verification/send-code": {
      "post": {
        "operationId": "VerificationController_sendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Верификационный код отправлен"
          },
          "429": {
            "description": "Слишком много запросов"
          }
        },
        "tags": ["Verification"]
      }
    },
    "/verification/verify-code": {
      "post": {
        "operationId": "VerificationController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Результат верификации",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many verification attempts"
          }
        },
        "tags": ["Verification"]
      }
    },
    "/reviews": {
      "post": {
        "operationId": "ReviewController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Reviews"]
      }
    },
    "/reviews/{id}/status": {
      "patch": {
        "operationId": "ReviewController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Reviews"]
      }
    },
    "/doctors": {
      "post": {
        "operationId": "DoctorController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно создан"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": ["Doctors"]
      }
    },
    "/doctors/{id}/schedule": {
      "get": {
        "description": "Полная информация о враче вместе с его расписанием",
        "operationId": "DoctorController_getDoctorSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID врача",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о враче и его расписании"
          },
          "404": {
            "description": "Врач не найден"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Получить врача с расписанием",
        "tags": ["Doctors"]
      }
    },
    "/vet-clinics": {
      "post": {
        "operationId": "VetClinicController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVetClinicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": ["Vet Clinics"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/doctors/{doctorId}/schedules": {
      "post": {
        "description": "Доступно только для пользователей с ролью VET_CLINIC своей клиники",
        "operationId": "DoctorScheduleController_create",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "ID врача",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Данные для создания расписания",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Запись в расписании успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorSchedule"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Врач не найден"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Создать запись в расписании врача",
        "tags": ["Расписание врачей"]
      },
      "get": {
        "description": "Получение всего расписания конкретного врача",
        "operationId": "DoctorScheduleController_findAll",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "ID врача",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список записей в расписании",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorSchedule"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Врач не найден"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Получить расписание врача",
        "tags": ["Расписание врачей"]
      }
    },
    "/doctors/{doctorId}/schedules/{id}": {
      "put": {
        "description": "Доступно только для пользователей с ролью VET_CLINIC своей клиники",
        "operationId": "DoctorScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID записи в расписании",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "ID врача",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Новый статус доступности",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoctorScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус доступности обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorSchedule"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Запись в расписании не найдена"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Обновить статус доступности врача",
        "tags": ["Расписание врачей"]
      }
    }
  },
  "info": {
    "title": "VetGuide API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID пользователя, для которого меняется роль",
            "example": 1
          },
          "newRole": {
            "type": "string",
            "description": "Новая роль пользователя",
            "enum": ["USER", "VET_CLINIC", "SUPER_ADMIN"],
            "example": "VET_CLINIC"
          }
        },
        "required": ["userId", "newRole"]
      },
      "SafeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "phone": {
            "type": "string",
            "example": "+79123456789"
          },
          "isVerified": {
            "type": "boolean",
            "default": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "phone", "isVerified", "createdAt", "updatedAt"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79123456789"
          },
          "fullName": {
            "type": "string",
            "example": "Иванов Иван Иванович"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "pasSwor!d1"
          },
          "confirmPassword": {
            "type": "string",
            "description": "Должен совпадать с паролем",
            "example": "pasSwor!d1"
          }
        },
        "required": ["phone", "fullName", "password", "confirmPassword"]
      },
      "SendCodeDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79123456789"
          }
        },
        "required": ["phone"]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79123456789"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["phone", "code"]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "doctorId": {
            "type": "number"
          }
        },
        "required": ["title", "description", "rating", "doctorId"]
      },
      "UpdateReviewStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Обновление статуса определенного отзыва",
            "example": "PENDING | VERIFIED"
          }
        },
        "required": ["status"]
      },
      "CreateDoctorDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "clinicId": {
            "type": "number"
          }
        },
        "required": ["fullName", "clinicId"]
      },
      "CreateVetClinicDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79123456789"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["phone", "password"]
      },
      "WeekDay": {
        "type": "string",
        "enum": [
          "Понедельник",
          "Вторник",
          "Среда",
          "Четверг",
          "Пятница",
          "Суббота",
          "Воскресенье"
        ],
        "description": "День недели"
      },
      "CreateDoctorScheduleDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "description": "День недели",
            "example": "Понедельник",
            "allOf": [
              {
                "$ref": "#/components/schemas/WeekDay"
              }
            ]
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Доступен ли врач в этот день",
            "example": true
          }
        },
        "required": ["dayOfWeek", "isAvailable"]
      },
      "Doctor": {
        "type": "object",
        "properties": {}
      },
      "DoctorSchedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Уникальный идентификатор записи",
            "example": 1
          },
          "dayOfWeek": {
            "description": "День недели",
            "example": "Понедельник",
            "allOf": [
              {
                "$ref": "#/components/schemas/WeekDay"
              }
            ]
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Доступен ли врач в этот день",
            "example": true,
            "default": true
          },
          "doctor": {
            "description": "Связанный врач",
            "allOf": [
              {
                "$ref": "#/components/schemas/Doctor"
              }
            ]
          }
        },
        "required": ["id", "dayOfWeek", "isAvailable", "doctor"]
      },
      "UpdateDoctorScheduleDto": {
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean",
            "description": "Статус доступности врача",
            "example": false
          }
        },
        "required": ["isAvailable"]
      }
    }
  }
}